Move analysis scripts to docs/ and update utils
- Relocate analyze_mbpp_library_usage.py, explore_humaneval.py,
  extract_mbpp_imports.py, test_mbpp_with_imports.py, and
  inspect_mbpp_import_test.ipynb to docs/
- Update common/utils.py and phase7_3 instruct baseline runner

+++ phase7_3_instruct_baseline/instruct_baseline_runner.py
@@ -238,7 +238,8 @@ class InstructBaselineRunner:
             )
             
             # Save as simple numpy array (matching Phase 3.5 format)
-            np.savez_compressed(save_path, layer_activations.clone().cpu().numpy())
+            # Convert to float32 before numpy conversion (BFloat16 not supported by numpy)
+            np.savez_compressed(save_path, layer_activations.clone().cpu().float().numpy())
     
     def _setup_output_directories(self) -> Path:
         """Create output directory structure and return output path."""
